More -05 changes
o  Adding a brief text on modifying the
session that is aligned with draft-ietf-mmusic-rid.
o  Improved text around simulcast stream identification (as opposed
to the simulcast stream itself) to consistently use the acronym
SCID and defined that in the Terminology section.
o  Improved IANA registration text with a reference to -mux-attributes
(since I specify MUX Category in the registration).
o  Removed unused reference to draft-ietf-payload-flexible-fec-
scheme.
o  More editorial improvements and corrections.