Additional tweakin of the changes.
Updated date, clarifying REQ 1-2, that this assumes knowledge of
signalling info, changing what corresponds differences of PTs and RIDs,
to allow also for other attributes, like a=imageattr by not restricting
it to those changes, removing PT duplication as motivation for not
allowing same SCID in multiple places on a a=simulcast line.

Corresponding text file included.