-09 as submitted
o  Changed SCID to rid-id, to align with ietf-draft-mmusic-rid

naming.
   o  Changed Overview to be based on examples and shortened
it.
   o  Changed semantics of initially paused rid-id in modified SDP

offers from requiring it to follow actual RFC 7728 pause state to

an informational offerer's opinion at the time of offer creation,

not in any way overriding or amending RFC 7728 signaling.
   o  Replaced
text on ignoring all but the first of multiple
      "a=simulcast" lines
in a media description with mandating that at
      most one
"a=simulcast" line is included.
   o  Clarified with a note that, for
the case it is clear from the SDP
      that RTP PT uniquely maps to
RtpStreamId, an RTP receiver can use
      RTP PT to relate simulcast
streams.
   o  Moved Section 4 Requirements to become Appendix A.
   o
Editorial corrections and clarifications.